MAP-5 upgrades codebase to be compliant with phpstan level 2

+++ app/Http/Controllers/API/NotificationController.php
@@ -1,10 +1,9 @@
-<?php
+<?php declare(strict_types = 1);
 
 namespace App\Http\Controllers\API;
 
 use App\Http\Controllers\Controller;
 use App\Models\User;
-use Illuminate\Http\Request;
 
 use App\Http\Requests\API\Notification\UpdateNotificationRequest;
 use App\Http\Requests\API\Notification\DeleteNotificationRequest;
@@ -21,19 +20,21 @@ class NotificationController extends Controller
     /**
      * Display a listing of the resource.
      *
-     * * @return \Illuminate\Http\Response
+     * * @return \Illuminate\Http\JsonResponse
      */
     public function index()
     {
         $user = Auth::user();
-        return response()->json( $user->notifications, 200 );
+        if($user instanceof User) {
+            return response()->json( $user->notifications, 200 );
+        }
     }
 
     /**
      * Find the specified resource by primary key.
      *
      * @param  int  $id
-     * @return \Illuminate\Http\Response
+     * @return \Illuminate\Http\JsonResponse
      */
     public function find($id)
     {
@@ -41,48 +42,60 @@ class NotificationController extends Controller
             return response()->json( [], 404 );
         }
         $user = Auth::user();
-        $notification = $user->notifications->where('id', $id)->first();
-       
-        if (empty($notification)) {
-            return response()->json( [], 404 ); 
+
+        if($user instanceof User) {
+            $notification = $user->notifications->where('id', $id)->first();
+               
+            if (empty($notification)) {
+                return response()->json( [], 404 ); 
+            }
+            return response()->json( $notification, 200 );
         }
-        return response()->json( $notification, 200 );
-        
+
+        return response()->json( [], 404 );
     }
 
     /**
      * Update the specified resource in storage.
      *
-     * @param  App\Http\Requests\API\Notification\UpdateNotificationRequest  $request
-     * @return \Illuminate\Http\Response
+     * @param  \App\Http\Requests\API\Notification\UpdateNotificationRequest  $request
+     * @return \Illuminate\Http\JsonResponse
      */
     public function update(UpdateNotificationRequest $request)
     {
         $user = Auth::user();
-        $notification = $user->notifications->where('id', $request->id)->first();
-        $notification->read_at = $request->read_at;
-        $notification->save();
 
-        return response()->json( $notification, 200 );
+        if($user instanceof User) {
+            $notification = $user->notifications->where('id', $request->id)->first();
+            $notification->read_at = $request->read_at;
+            $notification->save();
+
+            return response()->json( $notification, 200 );
+        }
+        return response()->json( [], 404 );
     }
 
     /**
      * Remove the specified resource from storage.
      *
-     * @param  App\Http\Requests\API\Notification\DeleteNotificationRequest  $request
-     * @return \Illuminate\Http\Response
+     * @param  \App\Http\Requests\API\Notification\DeleteNotificationRequest  $request
+     * @return \Illuminate\Http\JsonResponse
      */
     public function delete(DeleteNotificationRequest $request)
     {
         $user = Auth::user();
-        $notification = $user->notifications->where('id', $request->id)->first();
-        
-        if (empty($notification)) {
-            return response()->json( [], 404 ); 
-        }
-        
-        $notification->delete();
 
-        return response()->json( [], 204 );
+        if($user instanceof User) {
+            $notification = $user->notifications->where('id', $request->id)->first();
+            
+            if (empty($notification)) {
+                return response()->json( [], 404 ); 
+            }
+            
+            $notification->delete();
+
+            return response()->json( [], 204 );
+        }
+        return response()->json( [], 404 );
     }
 }
